move graphdino specific utility data functions to graphdino/
