+++ deep_neuronmorpho/data/__init__.py
@@ -1,4 +1,3 @@
 from .augmentations import augment_graph
 from .datasets import NeuronDataset
 from .process_swc import SWCData
-from .utils import compute_laplacian_eigenvectors
