+++ deep_neuronmorpho/graphdino/dataset.py
@@ -11,18 +11,18 @@ import torch
 from omegaconf import DictConfig
 from torch.utils.data import Dataset
 
-from ..data.utils import (
+from .augmentations import (
+    jitter_node_positions,
+    rotate_node_positions,
+    translate_all_nodes,
+)
+from .data_utils import (
     compute_path_lengths,
     drop_random_branch,
     find_leaf_nodes,
     neighbors_to_adjacency_torch,
     remap_neighbors,
     subsample_graph,
-)
-from .augmentations import (
-    jitter_node_positions,
-    rotate_node_positions,
-    translate_all_nodes,
 )
